+++ src/hosted-fields/internal/normalize-card-type.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var removeIgnorableCharacters = require('./remove-ignorable-characters');
+
+module.exports = function normalizeCardType(type) {
+  return removeIgnorableCharacters(type).toLowerCase();
+};