+++ src/hosted-fields/internal/remove-ignorable-characters.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = function removeIgnorableCharacters(str) {
+  if (str) {
+    return str.replace(/[-\s]/g, '');
+  }
+
+  return '';
+};